+++ lib/domain/bloc/task_detail_screen/task_detail_screen_state.dart
