+++ lib/presentation/screens/all_tasks/widgets/custom_slivertobox_adapter.dart
@@ -3,7 +3,7 @@ import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
 import 'package:todo/l10n/locale_keys.g.dart';
-import 'package:todo/bloc/all_tasks_screen/all_tasks_screen_bloc.dart';
+import 'package:todo/domain/bloc/all_tasks_screen/all_tasks_screen_bloc.dart';
 import 'package:todo/constants.dart' as Constants;
 
 class CustomSliverToBoxAdapter extends StatelessWidget {
