+++ lib/presentation/screens/all_tasks/widgets/task.dart
@@ -1,14 +1,14 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import 'package:todo/bloc/all_tasks_screen/all_tasks_screen_bloc.dart';
+import 'package:todo/domain/bloc/all_tasks_screen/all_tasks_screen_bloc.dart';
 import 'package:todo/constants.dart' as Constants;
 import 'package:todo/data/models/task_model.dart';
 import 'package:todo/presentation/widgets/svg.dart';
 import 'package:todo/helpers/format_date.dart';
 import 'package:todo/presentation/screens/task_detail/task_detail.dart';
-import '../../../../bloc/task_detail_screen/task_detail_screen_bloc.dart';
-import '../../../../navigation/tasks_router_delegate.dart';
+import 'package:todo/domain/bloc/task_detail_screen/task_detail_screen_bloc.dart';
+import 'package:todo/navigation/tasks_router_delegate.dart';
 import 'check_button.dart';
 
 class Task extends StatelessWidget {
