+++ lib/presentation/screens/all_tasks/widgets/visibility_button.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import 'package:todo/bloc/all_tasks_screen/all_tasks_screen_bloc.dart';
+import 'package:todo/domain/bloc/all_tasks_screen/all_tasks_screen_bloc.dart';
 import 'package:todo/constants.dart' as Constants;
 import 'package:todo/presentation/widgets/svg.dart';
 import 'package:todo/presentation/models/tasks_filter.dart';
