+++ lib/presentation/screens/task_detail/task_detail.dart
@@ -6,7 +6,7 @@ import 'package:device_info_plus/device_info_plus.dart';
 import 'package:todo/l10n/locale_keys.g.dart';
 import 'package:todo/constants.dart' as Constants;
 import 'package:todo/data/models/task_model.dart';
-import 'package:todo/bloc/task_detail_screen/task_detail_screen_bloc.dart';
+import 'package:todo/domain/bloc/task_detail_screen/task_detail_screen_bloc.dart';
 import 'package:todo/helpers/format_date.dart';
 import 'package:todo/presentation/screens/task_detail/widgets/material_textfield.dart';
 import 'package:todo/presentation/screens/task_detail/widgets/delete_button/delete_button.dart';
@@ -69,7 +69,7 @@ class TaskDetailScreenContent extends StatelessWidget {
             padding: const EdgeInsets.only(right: 8.0),
             child: Center(
               child: TextButton(
-                onPressed: () async {
+                onPressed: () {
                   bloc.add(const EditAcceptedEvent());
                   Navigator.pop(context);
                 },
